add open card open state
